反射

+++ src/cn/jxy/model/Product.java
@@ -1,5 +1,6 @@
 package cn.jxy.model;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 public class Product {
@@ -11,7 +12,7 @@ public class Product {
 //    date                timestamp default CURRENT_TIMESTAMP,*/
     private Integer id;
     private String name;
-    private Double price;
+    private BigDecimal price;
     private String remark;
     private Date date;
     private String pic;
@@ -21,7 +22,7 @@ public class Product {
         return "Product{" + "id=" + id + ", name='" + name + '\'' + ", price=" + price + ", remark='" + remark + '\'' + ", date=" + date + ", pic='" + pic + '\'' + '}';
     }
 
-    public Product(String name, Double price, String remark) {
+    public Product(String name, BigDecimal price, String remark) {
         this.name = name;
         this.price = price;
         this.remark = remark;
@@ -47,13 +48,9 @@ public class Product {
         this.name = name;
     }
 
-    public Double getPrice() {
-        return price;
-    }
+    public BigDecimal getPrice() { return price; }
 
-    public void setPrice(Double price) {
-        this.price = price;
-    }
+    public void setPrice(BigDecimal price) { this.price = price; }
 
     public String getRemark() {
         return remark;
